Luan update 27-10

+++ services/order-service/src/main/java/com/smartdelivery/order/dto/PriceQuoteSnapshot.java
@@ -1,4 +1,3 @@
-// dto/PriceQuoteSnapshot.java
 package com.smartdelivery.order.dto;
 
 import lombok.Builder; import lombok.Data;
@@ -6,7 +5,7 @@ import java.math.BigDecimal; import java.time.OffsetDateTime;
 
 @Data @Builder
 public class PriceQuoteSnapshot {
-    private BigDecimal distanceKm;
+    private Double distanceKm;
     private Integer travelTimeMin;
     private BigDecimal priceAmount;
     private String priceCurrency;
